Separate image storage per mechanism (sixel, iip)
